Move the ImgTransformation to common
