+++ include/depthai-shared/common/RawImgTransformation.hpp
